+++ flutter_application_1/lib/Widgets/page_view_demo.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:hexcolor/hexcolor.dart';
 import 'package:smooth_page_indicator/smooth_page_indicator.dart';
 import '../Screens/Home.dart';
 import '../Screens/Preview.dart';
